+++ Ray-Tracing/light_source.cpp
@@ -1,3 +1,3 @@
 #include "light_source.h"
 
-const double LightSource::baseShining = 0.1;
+const long double LightSource::baseShining = 0.1;
